Added ProgressBar, Removed RadioButton, Etc
Removed the radioButton control because of compatibility issues.

Added progressBar controls and relating methods. The value of this
control can be set as an int32 or a string formatted as '-10' or '+10'
for examples.

Added tZComponent.equals() methods.

Fixed an issue with setting the readOnly property.

ComboBox and listbox now accept tStringArray rather than using the pipe
delimiter.