Added Passwords, Image URLs, Methods, and More!
eZForm now requires ProSocks. This went without question, as various
functions, such as getPage() and rewriteFile() would cause access
violations when used in event handlers. See initial post of the eZForm
thread for more details.

Changed various variable names to be consistent with their methods.

Added the password property and methods to edits and memos. An example
has been added to the initial post of the eZForm thread.

Added the handle property and methods to all components.

Added the src property and methods to images. An example has been added
to the initial post of the eZForm thread.

Removed tZForm.addImageURL(). If a URL, requiring the preceding http://,
is passed to tZForm.addImage(), it'll automatically download and add the
image.