Added Tab Elements, Parent Method, Etc
Fixed automatic sizing issues.

Added the tab and tabList elements.

Added the eZForm.addTab() method. If an ID is passed, for example iTabs,
the tabs created will be given the IDs iTabs0, iTabs1, iTabs2, etc. An
example has been added to the initial post of the eZForm thread.

Added the select property and methods to tabList elements. These will
return or set the selected tab in a tabList.

Added the parent property and methods to all elements. These will return
or set the parent of an element. An example has been added to the
initial post of the eZForm thread.