Bug Fixes, Auto-Sizing Images, Etc
Images now automatically size themselves if the dimensions parameter is
set to [0,0] when using eZForm.addImage().

Changing the visible properties of multiple elements will now work as
expected. Previously, hiding or showing more then a single element in
one thread would seemingly ignore the rest.

eZForm.show() is now a function and will return the current form. This
is useful for setting a custom dimensions and changing properties on the
fly, for example:
eZForm.show().height(240);